python_code/api/requirements.txt
pandas==2.0.3
python-dotenv==1.0.1
openai==1.50.2
pinecone==5.4.0
runpod==1.7.1